Move enums to component folders
