+++ src/main/java/frc/robot/Robot.java
@@ -10,7 +10,7 @@ import frc.robot.constants.field_constants.HomeField;
 import frc.robot.constants.robots.RobotA;
 import frc.robot.vision.Limelight;
 import frc.robot.subsystems.DrivetrainInterface;
-import frc.robot.subsystems.LED;
+import frc.robot.subsystems.LED.LED;
 
 public class Robot extends TimedRobot {
   private Command autoCommand;
