+++ src/main/java/frc/robot/motion_profiling/FollowPath.java
@@ -7,7 +7,7 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj.trajectory.Trajectory;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import edu.wpi.first.wpilibj2.command.RamseteCommand;
-import frc.robot.enums.AutoPath;
+import frc.robot.motion_profiling.AutoPath;
 
 import static frc.robot.Robot.drivetrain;
 import static frc.robot.Robot.robotConstants;
