+++ src/main/java/frc/robot/subsystems/LED/LED.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.LED;
 
 import java.util.Random;
 
@@ -8,7 +8,7 @@ import edu.wpi.first.wpilibj.Spark;
 import edu.wpi.first.wpilibj.DriverStation.Alliance;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Robot;
-import frc.robot.enums.LEDColor;
+import frc.robot.subsystems.LED.LEDColor;
 
 public class LED extends SubsystemBase {
   private static final double BLINK_TIME = 0.2;
