+++ src/main/java/frc/robot/vision/FollowTarget.java
@@ -5,7 +5,7 @@ import edu.wpi.first.wpilibj.controller.PIDController;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Robot;
-import frc.robot.enums.Target;
+import frc.robot.vision.Target;
 import frc.robot.utils.PIDSettings;
 import java.util.function.BiConsumer;
