+++ src/main/java/frc/robot/vision/LedMode.java
@@ -1,4 +1,4 @@
-package frc.robot.enums;
+package frc.robot.vision;
 
   public enum LedMode {
     defaultPipeline(0), off(1), blink(2), on(3);
